FirmwareUpdate: Move a few things around based on suggestions
Moved the firmware changelog next to the latest firmware, reworded a few
strings, and moved the current firmware version out of the radio group.
